+++ react/components/OrderItems.js
@@ -1,51 +1,17 @@
 import React from 'react'
-import { arrayOf, number, shape, string } from 'prop-types'
-import { prop, path } from 'ramda'
-import { Link } from 'vtex.render-runtime'
-import ProductPrice from 'vtex.store-components/ProductPrice'
+import { arrayOf } from 'prop-types'
 
-import { isParentItem, separateParentChildren, sumTotalPricePerUnit } from '../utils/attachments'
+import { orderItemType } from './propTypes'
+import OrderItem from './OrderItem'
 
-const OrderItem = ({ item, index, assemblyOptions }) => (
-  <li
-    key={prop('uniqueId', item)}
-    className="vtex-rebuy__item flex justify-between pt3 bb b--muted-4"
-  >
-    <div className="vtex-rebuy__item-column flex-shrink-0 mr5">
-      <Link className="link c-muted-1" to={path(['detailUrl'], item)}>
-        <img
-          height="64"
-          width="64"
-          className="vtex-rebuy__item-img"
-          src={prop('imageUrl', item)}
-        />
-      </Link>
-    </div>
-    <div className="vtex-rebuy__item-column flex-auto">
-      <div className="vtex-rebuy__item-details flex justify-center justify-between-ns items-center-ns flex-column flex-row-ns h-100 mb2-s">
-        <Link className="link t-body t-heading-5-ns c-on-base" to={path(['detailUrl'], item)}>
-          <span className="vtex-rebuy__item-details">{`${prop(
-            'quantity',
-            item
-          )}x ${prop('name', item)}`}</span>
-        </Link>
-        <span className="vtex-rebuy__item-price ">
-          <ProductPrice
-            sellingPriceClass="c-on-base t-body t-heading-5-ns b-s"
-            sellingPrice={sumTotalPricePerUnit(assemblyOptions, index, item) / 100}
-            showListPrice={false}
-            showLabels={false}
-          />
-        </span>
-      </div>
-    </div>
-  </li>
-)
+import { isParentItem, separateParentChildren } from '../utils/attachments'
+
+import rebuy from '../rebuy.css'
 
 const OrderItems = ({ items = [] }) => {
   const [_, assemblyOptions] = separateParentChildren(items)
   return (
-    <ul className="vtex-rebuy__items list ma0 pa0">
+    <ul className={`${rebuy.items} list ma0 pa0`}>
       {items.map((item, index) => isParentItem(item) ? (
         <OrderItem {...{ item, assemblyOptions, index}} />
       ) : null)}
@@ -54,19 +20,7 @@ const OrderItems = ({ items = [] }) => {
 }
 
 OrderItems.propTypes = {
-  items: arrayOf(
-    shape({
-      uniqueId: string.isRequired,
-      imageUrl: string.isRequired,
-      quantity: number.isRequired,
-      sellingPrice: number.isRequired,
-      parentItemIndex: number,
-      parentAssemblyBinding: string,
-      id: string.isRequired,
-      name: string.isRequired,
-      detailUrl: string,
-    })
-  )
+  items: arrayOf(orderItemType),
 }
 
 export default OrderItems